% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{resC3CO-RK29}
\alias{resC3CO-RK29}
\title{resC3CO-RK29}
\format{A c3coFit object from package c3co
\describe{
  \item{bkp}{breakpoints}
  \item{segDat}{Segmented data}
  \item{fit}{Results of c3co for each number of feature}
  ...
}}
\description{
dataset containing the output of C3CO method on patient RK29 samples from GSE47077 dataset
}
