% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{segDat-RK29}
\alias{segDat-RK29}
\title{segDat-RK29}
\format{A list with 4 components
\describe{
  \item{bkp}{List of breakpoints}
  \item{Y1}{Segmented Minor copy number}
  \item{Y2}{Segmented Major copy number}
  \item{Y}{Segmented Total copy number}
  ...
}}
\description{
dataset containing the segmentation of data for patient RK29
}

