% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{segment,GSE22928,MM,LAU149PBL}
\alias{segment,GSE22928,MM,LAU149PBL}
\title{segment,GSE22928,MM,LAU149PBL}
\format{A list with 5 components:
\describe{
  \item{data}{PairedPSCNData}
  \item{output}{PairedPSCNSegments}
  \item{tcnSegRows}{Segments of TCN}
  \item{dhSegRows}{Segments of dh}
  \item{params}{parameters for segmentation}
  ...
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE22928}
}
\description{
dataset containing the output of PSCBS package
}

